Simulated the dynamic performance of the bicycle model in reference configuration. The model was imported from BicycleModel in model.py.
However, I found the forcing_full matrix has uas and uads of auxiliary
speeds in it, which mean a problem of substitution may exist in KaneMethod
of mecha, although the forcing matrix is right after replacing uas and
uads with zeros. Overall, the forcing matrix issue should go to Gilbert for
an answer.